refactor to shared cloudflare upload stuff
